service: add File(Metadata)Service stubs + refactor
